First attempt at real fix.
We bind PHP callables in to the engine directly,
using userdata. This means we don't need to pass
through a callback index, nor maintain a static
list of registered callbacks. Instead, we simply
hold a list of callbacks registered per lua object,
and clean them up when the object is freed. This
achieves a fix of the memory leaks without need
for an obscure method on the LUA extension.

This commit has broken some tests. More work needed.